basic function finished

+++ src/main/java/com/weiyuscode/wechatpush/entity/Weather.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class Weather {
-    Float tempMin;
-    Float tempMax;
-    Float tempFeel;
+    Integer tempMin;
+    Integer tempMax;
+    Integer tempFeel;
     String weatherDetails;
 }
