+++ src/main/java/com/weiyuscode/wechatpush/service/impl/WeatherServiceImpl.java
@@ -17,16 +17,16 @@ public class WeatherServiceImpl implements WeatherService {
     public void renewWeather(JSONObject weatherInfo) {
         float kelvinsOffset = (float) 273.15;
         JSONObject main = weatherInfo.getJSONObject("main");
-        Float tempMin = main.getFloat("temp_min") - kelvinsOffset;
-        Float tempMax = main.getFloat("temp_max") - kelvinsOffset;
-        Float tempFeel = main.getFloat("feels_like") - kelvinsOffset;
+        float tempMin = main.getFloat("temp_min") - kelvinsOffset;
+        float tempMax = main.getFloat("temp_max") - kelvinsOffset;
+        float tempFeel = main.getFloat("feels_like") - kelvinsOffset;
         String weatherDetail = weatherInfo.getJSONArray("weather").getJSONObject(0).getString("description");
 
         Weather newWeather = new Weather();
         newWeather.setWeatherDetails(weatherDetail);
-        newWeather.setTempFeel(tempFeel);
-        newWeather.setTempMax(tempMax);
-        newWeather.setTempMin(tempMin);
+        newWeather.setTempFeel((int) tempFeel);
+        newWeather.setTempMax((int) tempMax);
+        newWeather.setTempMin((int) tempMin);
         todayWeather = newWeather;
         System.out.println("get weather info: " + todayWeather);
     }
